@@ -1,9 +1,9 @@
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
-import zhCommon from "i18n/zh_cn/common"
-import enCommon from "i18n/en/common";
-import enFooter from "i18n/en/footer";
+import zhCommon from "./zh_cn/common"
+import enCommon from "./en/common";
+import enFooter from "./en/footer";
 import zhFooter from "./zh_cn/footer";
 import zhRules from "./zh_cn/rules";
 import enRules from "./en/rules";
