@@ -2,7 +2,7 @@ import { Outlet } from 'react-router-dom';
 import {useLocation} from "react-router";
 
 import { OngoingWrapper } from "./style";
-import MyPrediction from "pages/my_prediction";
+import MyPrediction from "../my_prediction";
 
 const Ongoing = () => {
     const location = useLocation();
