@@ -3,11 +3,11 @@ import styled from "styled-components";
 import {LeftOutlined, RightOutlined} from "@ant-design/icons";
 import {Carousel, Spin} from "antd";
 
-import CoinCard from "components/coin_card";
-import {ApiContext, Prediction} from "App";
-import ContentHeader from "components/content_header";
-import {predictionSort} from "utils/prediction-sort";
 import {formatHumanNumber} from "../../../utils/format";
+import {ApiContext, Prediction} from "../../../App";
+import CoinCard from "../../../components/coin_card";
+import {predictionSort} from "../../../utils/prediction-sort";
+import ContentHeader from "../../../components/content_header";
 
 
 const UpcomingFluctuations = () => {
